+++ frontend/src/app/[locale]/api-dashboard/layout.tsx
@@ -1,0 +1,25 @@
+import { Metadata } from "next";
+import { LayoutProps } from "src/types/generalTypes";
+import { LocalizedPageProps } from "src/types/intl";
+
+import { getTranslations } from "next-intl/server";
+
+import { AuthenticationGate } from "src/components/user/AuthenticationGate";
+
+export async function generateMetadata({ params }: LocalizedPageProps) {
+  const { locale } = await params;
+  const t = await getTranslations({ locale, namespace: "ApiDashboard" });
+  const meta: Metadata = {
+    title: t("pageTitle"),
+    description: t("metaDescription"),
+  };
+  return meta;
+}
+
+export default function ApiDashboardLayout({ children }: LayoutProps) {
+  return (
+    <>
+      <AuthenticationGate>{children}</AuthenticationGate>
+    </>
+  );
+}
